Update JSONRPC and REST validators
