+++ qos/cosmos/rpctype_cosmos.go
@@ -1,6 +1,55 @@
 package cosmos
 
-import "strings"
+import (
+	"strings"
+
+	sharedtypes "github.com/pokt-network/poktroll/x/shared/types"
+)
+
+// ------------------------------------------------------------------------------------------------
+// REST API Type Detection - Consolidated Logic
+// ------------------------------------------------------------------------------------------------
+
+// determineRESTRPCType determines the specific RPC type for REST requests
+func determineRESTRPCType(path string) sharedtypes.RPCType {
+	// CometBFT REST-style endpoints should be tagged as COMET_BFT
+	if isCometBftRpc(path) {
+		return sharedtypes.RPCType_COMET_BFT
+	}
+
+	// Everything else is regular REST
+	return sharedtypes.RPCType_REST
+}
+
+// isValidRESTPath validates that the path matches REST API patterns
+func isValidRESTPath(path string) bool {
+	// CosmosSDK REST API patterns
+	if isCosmosRestAPI(path) {
+		return true
+	}
+
+	// CometBFT REST-style endpoints
+	if isCometBftRpc(path) {
+		return true
+	}
+
+	// Allow some additional common REST patterns
+	additionalValidPaths := []string{
+		"/txs",          // Legacy transaction endpoint
+		"/tx",           // Transaction endpoints
+		"/node_info",    // Node information
+		"/syncing",      // Sync status
+		"/latest_block", // Latest block info
+	}
+
+	for _, validPath := range additionalValidPaths {
+		if path == validPath || (len(path) > len(validPath) && path[:len(validPath)+1] == validPath+"/") {
+			return true
+		}
+	}
+
+	return false
+}
 
 // ------------------------------------------------------------------------------------------------
 // Cosmos SDK RPC Type Detection
